+++ examples/llamaIndexSample.js
@@ -7,7 +7,7 @@ setupMonocle(
     new BatchSpanProcessor(
       new ConsoleSpanExporter(),
       config = {
-        scheduledDelayMillis: 20
+        scheduledDelayMillis: 1
       })
   ]
 )
