+++ src/metamodel/maps/llamaindex/llamaindex_methods.js
@@ -17,7 +17,7 @@ exports.config = [
         "method": "chat",
         "spanName": "llamaindex.openai_chat",
         "output_processor": [
-           require("./attributes/llamaindex_attributes.js").config
+           require("./llamaindex_attributes.js").config
         ]
 
     },
