+++ src/metamodel/maps/openai/openai_attributes.js
